requirements.txt
# 数据处理
pandas>=1.3.0
numpy>=1.21.0
openpyxl>=3.0.7

# 配置文件
pyyaml>=5.4.1

# 文本处理
emoji>=1.7.0
tqdm>=4.62.0
regex>=2023.0.0

# 情感分析相关
snownlp>=0.12.3
transformers>=4.11.0
torch>=1.9.0
hanlp>=2.1.0

# 深度学习模型
scikit-learn>=0.24.2 